Chore: Rename folder from 'style' to 'styles'
