+++ src/components/MenuInfo/components/ButtonRemoveMember/index.tsx
@@ -1,6 +1,6 @@
 import { UsergroupDeleteOutlined } from '@ant-design/icons'
-import '../../style/style.css'
-import { ButtonDefaltStyle } from '../../style/style'
+import '../../styles/style.css'
+import { ButtonDefaltStyle } from '../../styles/style'
 import { ButtonHTMLAttributes } from 'react'
 
 interface ButtonRemMemberProps extends ButtonHTMLAttributes<HTMLButtonElement> {
