+++ src/components/MenuInfo/styles/style.css
