+++ src/components/MenuInfo/styles/style.tsx
